Move synchronous flash implementation for high cycle area to synchronous hal_st and add WithIrqHandler
